review fixes: decimal targets, atomic uploads, validate-then-apply, restart and flash guards
Second review of the legacy ESP fixes:

- /setvalve: the web UI posts the target field as a string, and the
  strtol-only parser rejected "50.5", "50.0" and "5e1", which the base
  branch (ArduinoJson as<uint8_t>) truncated to 50. Numeric strings may
  now be decimal (strtod, whole string, finite, in range), truncated as
  JSON numbers are. Hex, inf, nan and junk stay rejected.
- /setvalve validates the valve, target, ctrlValue, ctrlTarget and
  ctrlDynOffs first and changes nothing when one is invalid, instead of
  applying the PI values and answering 400. /cmd vCalib/vAssembly with an
  invalid valve now answers 400 instead of a silent 200.
- Temperature offsets: non-finite input keeps the stored offset; the
  offset is clamped to +/-10.0 degrees on input and again where it is
  applied (so values stored by older firmware are bounded), and a valid
  reading can no longer wrap int16 or drop to the <= -500 "failed"
  sentinel range.
- UART: only the reply to the outstanding request clears appTimeOuts
  and stmFailed. A late reply or noise reset the count, so a STM that
  always answered after the timeout was never reset.
- STM flash guard: a stmFlashActive flag is set before the request's 1 s
  delay. startApp()/startMqtt() no longer create tasks while it is set
  (the delayed first start could start app_loop on UART2 next to the
  bootloader), and the flash is abandoned before erasing if a restart
  was scheduled concurrently.
- Restarts cannot cut off a STM flash: restartSystem() defers while the
  flash runs (the flasher restarts the ESP when done; /cmd reboot
  answers 409), ESP OTA uploads are refused with 409 while it runs, and
  an ESP image completed during a flash boots at the flasher's restart.
- Uploads: the finished .part file is renamed over the old file (LittleFS
  rename replaces atomically); the old file is removed only if the
  rename fails, then renamed again. The upload status byte is allocated
  before the file is opened, so recording a failure needs no memory and
  a failed allocation answers 500 instead of a false 200.

The Arduino-free parsing and offset code moves to valueCheck.{h,cpp}
with host tests in software_esp32/test/native (doctest, ASan/UBSan).

+++ software_esp32/include/stmApp.h
@@ -219,7 +219,7 @@ private:
   void app_web_cmd_check();
   void setSensorIndex(uint8_t valveIndex,char* sensor1,char* sensor2); 
   int16_t ConvertCF(int16_t cValue);
-  int16_t getTOffset(uint8_t tIdx);
+  int getTOffset(uint8_t tIdx);
   bool checkNewTarget();
 
   bool fastQueueMode;
